BioSecure-ID-System/requirements.txt
Flask
psycopg2-binary
face_recognition
opencv-python
numpy
python-dotenv
